requirements.txt
annotated-types==0.6.0
greenlet==3.0.1
numpy==1.24.4
pandas==2.0.3
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
tzdata==2023.3
